feat: reorganized the billing ui page
